Add FFI for Cactus model initialization and transcription
- Implemented cactus_init function for model initialization with error handling.
- Added cactus_transcribe function to process audio input and generate transcriptions.
- Introduced utility functions for handling spectrogram computation and normalization.
- Created CactusModelHandle structure to manage model state and mutex for thread safety.
- Added error handling and response formatting utilities for FFI interactions.
- Included necessary headers and dependencies for audio processing and JSON handling.
